requirements.txt
streamlit
openai
python-dotenv
chromadb
langchain
langchain-community
tiktoken
langchain-openai
pypdf2